Add BitClient class and implement registerClient

+++ lib/connection/serial_connection.h
@@ -5,7 +5,7 @@ struct SerialConnection : public Connection
 {
 public:
     SerialConnection();
-    void open(int baud_rate);
+    void open(unsigned long baud_rate);
     void close();
     bool send(const pb_msgdesc_t *field, const int msg_id, const void *msg_struct);
 };
